+++ start.sh
@@ -1,3 +1,7 @@
-#!/bin/bash
+#!/usr/bin/env bash
+
+BASEDIR=$(dirname "$0")
+
+cd "$BASEDIR" || exit 1
 
 npm start